tests/kms_concurrent: Use maximum number of planes that display/platform combination supports
There was an error in

commit 0ab05a51a059 ("tests/kms_concurrent: Test for maximum number of planes")

that calculates the maximum number of supported planes. The patch proposes to
test first if an atomic commit is successful with the selected number of planes.
In case of a failure, the number of planes is reduced by one assuming that the
failure was caused by bandwidth limit. The test loops at least as many iterations
as there are planes defined by the platform.

v2: Fix CI build failure (CI)

Signed-off-by: Mika Kahola <mika.kahola@intel.com>
Reviewed-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>