kms_flip: Skip interruptible testing for negative and hangs tests
Negative tests are API tests that will be rejected before we ever wait.
Repeating them expecting it to be interrupted is a forlorn hope.

Hang tests are expected to be blocked until they can be resolved by
hangcheck. Repeating those to see how a long wait responds when
interrupted is counter purpose, just a waste of time.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>