i915/suspend: Suspend once before taking forcewake to confirm suspend
If we take the user-forcewake and then exit, we find outselves in a
deadlock as the atexit handlers wait for our own forcewake to be
released. We could either not use the automatically attached exit
handler (__drm_open_driver), more carefully unwind the exit handler, or
simply avoid doing late requirement checks. This patch does the later by
testing for working suspend prior to taking the user-forcewake.

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/1637
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>