i915/gem_exec_alignment: Limit priority inversion warning
The premise of the test is that one client does not block the other, to
which we only need to verify that the ready client proceeds unhindered
by the abusive client. So we only need assert that the ready client
completes within a shorter time than the abusive client.

It would be nice that we could assert that under all conditions it never
takes more than a millisecond to submit a no-op; but that goal is a bit
too far for today.

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/1654
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>