kms_flip: Reduce hang tests to minimum duration
Hang tests are slow since we must wait for the kernel to detect the GPU
hang before it forcibly completes the execution and signals the fence.
The flip is not allowed before the fence is signaled and so must wait.
Our test is that the flip does eventually occur, we only need to do that
once, and let the power of repeated runs across many machines weed out
the corner cases.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>