intel-ci: Drop gem_exec_suspend@basic-S4-device from BAT
S4-devices is a debug only path, and while it can prove informative as
to the nature of suspend issues, being a debug only path it is not so
relevant towards the driver health. Relegate it to the shards and idle
runs, so we can shave 20s off BAT.

gem_exec_suspend@basic itself is just a test suspend, a sanitycheck to
ensure the test is functional, for manual debugging of the tests.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Martin Peres <martin.peres@linux.intel.com>
Cc: Petri Latvala <petri.latvala@intel.com>
Acked-by: Martin Peres <martin.peres@linux.intel.com>