tests/kms_fbcon_fbt: Make FBC wait_until_update() more reliable
We can't depend onto debugfs reads of the FBC status as compression
takes one idle frame to recompress and we could easily miss that.

Instead lets use the pipe CRC as it keeps the CRC of each frame so we
can compare each other until a blink in the FBCON causes it do change.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Reviewed-by: Imre Deak <imre.deak@intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>