i915/gem_tiled_fence_blits: Reduce working set size
Our goal with this gem_*_blits variant is to exercise fence reuse, and
since the fenced blits [of yore] can only fit into the fenceable region
of the mappable aperture, we can limit our working set to cause
thrashing of that region. This dramatically reduces the workload for big
machines that were trying to trash everything.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>