i915/gem_exec_alignment: Cap object setup time
Instantiating the objects can be surprisingly, worryingly slow; even on
big core LLC devices. Rather than worry about that and cause timeouts in
simply setting up the test, apply a limit to the setup time so that we
only use the amount of objects instantiated within the first 20s.

Due to the quadratic behaviour being exposed by the test, this may not
entirely eliminate the timeouts -- but at least the timeouts should then
be for the right reason.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>