perf: Verify rc6 works first
We want to verify that OA disables rc6 (so that its metrics are not
interrupted by the device sleeping), and once OA is finished, the device
is allowed to sleep again. If rc6 is broken, the test fails since rc6
never recovers -- but that is expected as rc6 is broken.

Skip the test rather than fail for an external issue; we know rc6 is
broken on ehl!

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/1042
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>