tests/kms_multipipe_modeset: Add test to validate max pipe configuration
Added test to validate max pipe configuration for a platform.
In the test, the reference CRC is collected first by doing modeset
on all the outputs individually.Then a simultaneous modeset is
done on all the outputs and the CRC is collected. This is compared
with the reference CRC.

If the number of outputs connected is less than the number of pipes
supported by the platform, then the test skips.

This test is added to verify if the max pipe configuration for a
given platform is working fine. Even though there are other tests
which test multipipe configuration, they test some other functionalities
as well together with multipipe. This is a stand alone test that
intends to only verify simultaneous modeset at the max pipe configuration.

v2: -Fix Typo in comment (Petri)
    -Use igt_require_pipe_crc() to prevent crash in non i915 case (Petri)
    -Print the number of outputs and pipes in igt_require() for
     debugging aid (Petri)

v3: -Change the subtest name (Petri)
    -Avoid multiple loops and remove the output_to_pipe variable,
     by making the output to pipe mapping fixed (Ankit)

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>
Acked-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>