i915/gem_exec_alignment: Directly compare naughty/normal execution times
If the kernel improves upon its O(N^2) alignment search, it can easily
fit a few GiB of buffers into the GTT within the expected minimum run
time. Instead of hardcoding the expected bad behaviour, our assertion is
that the naughty client does not hinder the normal client, so directly
compare the two execution times. The normal client should not be doing
any work, so we never expect it to take more than a few microseconds...

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>