i915/gem_exec_reloc: Trim flooding
Using a quadruplicing load factor and a 5s timeout, meant that we could
submit one more workload just before the timeout that would take around
20s for the kernel to process (just from the large number of objects and
relocations). As that exceeds the hangcheck, the spinner was being
killed even though we did not hit the deadlock. Use a doubling factor,
and aim for the worst case run to be less than 10s.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>