i915/gem_exec_params: Don't skip for invalid-bsd-ring
In the CI shards, we have a mixture of icl, some with vcs2, some without.
Any test that requires vcs2 will flip-flop between skips, so avoid such
requires where possible. Here, we know what the outcome should be if we
have vcs2 and if we do not, so assert we get the correct outcome.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>