kms_flip: Trim test durations for basic subtests [BAT]
Reduce the target runtime for the individual subtests as each may be
repeated multiple times accumulating a substantial fraction of the BAT
runtime. A 100 flips or so should be as good an indicator of basic
driver health as a few thousand, with the stress testing coming in the
non-basic testing.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Petri Latvala <petri.latvala@intel.com>