intel-ci: Leave gem_ctx_param to the shards
Checking the basic API behaviour has only minimal driver coverage
(essentially just the syscall validation paths) and does not provide
good indication of wider driver health, nor do we expect the results to
vary between machines (as we are only scratching the surface). There is
no advantage to running these early in BAT, leave it to the shards.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>