tests/kms_flip: Keep signal helpers enabled after a failed interruptible subtest
For the duration of all the interruptible subtests the signal helper is
assumed to stay running after each subtest, both in case the subtest
passes or aborts with a failure. So make sure we don't leave the signal
helper suspended in case igt_assert() aborts.

References: https://gitlab.freedesktop.org/drm/intel/issues/1883
Fixes: 60e8be7ccc72 ("tests/kms_flip: Retry test in case of a DP/HDMI link reset")
Cc: Uma Shankar <uma.shankar@intel.com>
Cc: Nidhi Gupta <nidhi1.gupta@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
Reviewed-by: Anshuman Gupta <anshuman.gupta@intel.com>