tests/fbcon: Don't close drm fd
For GEN9+ fbc_wait_until_update() fetches pipe CRCs to check if it is
back to fbcon and if cursor is blinking but to return to fbcon it was
closing the drm fd.

This code was only working because CRC functions only uses drm fd to
get the debugfs dir fd and igt_debugfs_dir() tries to fallback to the
default path in case drm fd == -1, anything different would cause the
tests to fail.

So here not closing the drm fd and still going back to fbcon, the
tricks to do it is implemented in restore_fbcon().

Not closing drm fd allowed us to remove setup_drm() and teardown_drm()
and only doing the setup and teardown once.

v2:
- fixed commit description

Reviewed-by: Matt Atwood <matthew.s.atwood@intel.com>
Cc: Matt Atwood <matthew.s.atwood@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>