lib/kms_chamelium: Wait for the sink to reconnect after an FSM DPMS-off
After Chamelium does an FSM signaling by deasserting its HPD, in
response to which the test's FSM handler disables the output with a
DPMS-off, we have to make sure that Chamelium has reasserted its HPD
before we re-enable the output with DPMS-on (for instance to avoid link
training errors, or enable the output in the wrong TypeC mode). To
ensure this wait for the connector state to become asserted.

On TypeC connectors with an enabled mode the IOM firmware will signal a
connected state (via a connect hotplug interrupt delivered to the
driver) in a deferred way only after the mode is disabled. So wait for
the connected state after DPMS-off.

Reported-and-tested-by: Kunal Joshi <kunal1.joshi@intel.com>
Cc: Hiler Arkadiusz <arkadiusz.hiler@intel.com>
Cc: Kunal Joshi <kunal1.joshi@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>