i915/gem_exec_balancer: Exercise inter-locked synchronous bonds
Add another bonded-runner variant for a pair of inter-locked
non-preemptible waiters (each batch waits to be cancelled by the other),
as we are trying to replicate the HW feature where a bonded pair is
locked on HW and cannot be preempted until all members of the bond have
executed (i.e. all must execute in lockstep).

The key difference with this variant is that both halves contain a
non-preemptible spin until the other half has executed.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>