i915/i915_pm_rps: Tighten fence generation for waitboost
We need the object we wait on to be clean, and free of stray fences as
the waitboosting only applies to i915_request that have not been
started. If we use an exclusive fence (from a write dependency) that
will strip off any foreign fences, allowing us to trick the driver into
behaving as we expect.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>