lib: Use generic names for APIs that handle uevents
The functions for handling uevents are named with "_hotplug" as suffix
such as igt_watch_hotplug(). Earlier hotplug was the only uevent that
was requested to be detected, but in fact, these APIs are generic and
can be used for detecting other uevents too.

Currently we have tests like kms_lease, kms_content_protection using
the uevent handling infrastructure for detecting uevents other than
hotplug.

This patch renames the functions and replace the "_hotplug" suffix
with "_uevents".

v2: Rebase
v3: Added changes in lib igt_chamelium.c, that was missed earlier.

Signed-off-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>
Suggested-by: Ramalingam C <ramalingam.c@intel.com>
Suggested-by: Hiler, Arkadiusz <arkadiusz.hiler@intel.com>
Reviewed-by: Ramalingam C <ramalingam.c@intel.com>
Acked-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>