lib/i915: Open primary sysfs for engine attributes
The engine attributes are only found on the primary sysfs node, not the
render nodes. If we are passed in a render node, we need to find the
corresponding primary node. We can do this by searching the character
devices for a primary minor index for the same exact device.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>