i915/gem_ctx_exec: Set a fast preempt timeout for nohangcheck
We want the test to run in a reasonable time, and verify that if
the hangcheck modparam is disabled we can still recover from a dead
context. The issue is that we've disabled preempt_timeout on tgl/rcs0 by
default, and so since we want this test to be fast, we need to re-enable
it for the test. Otherwise, we would have to wait several seconds for
the heartbeat to notice the hung engine, exceeding the limits imposed on
the test.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>