runner: Also generate igt@runner@aborted when aborting internally
If we can't kill the (main) test process, or when the test process
exits with IGT_EXIT_ABORT, we abort the execution. Pass that
information along to the other machinery that tracks whether we
aborted, thus also getting that information to the end user in the
form of the pseudo-result igt@runner@aborted.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Cc: Lukasz Fiedorowicz <lukasz.fiedorowicz@intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>