tests: Add feature_discovery
The main reason for adding this test is automatic tagging of the
machines for the purposes of CI and other automation systems.

The subtest are not doing much - they just check the presence of
interfaces and available hardware to make sure certain "feature" is
there and ready to be tested properly.

The subtests should only pass or skip.

Any failures or flip-floppers are reason for serious concern.

Cc: Martin Peres <martin.peres@intel.com>
Cc: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Lukasz Fiedorowicz <lukasz.fiedorowicz@intel.com>